teste funcionando eu acho

+++ src/main/java/br/unitins/dto/AcessorioDTO.java
@@ -1,14 +1,13 @@
 package br.unitins.dto;
 
-import java.math.BigDecimal;
-
 public record  AcessorioDTO ( String nome,
  String descricao,
- BigDecimal preco,
+ Double preco,
  String imagemUrl,
 String cor,
  String tamanho) {
-   
+
+
 
     /*public AcessorioDTO(String nome, String descricao, BigDecimal preco, String imagemUrl, String cor, String tamanho) {
         this.nome = nome;
